Improved Prometheus export with technical data / API Documentation improvement.
Prometheus now exports statistics including all requests performed to Consul
agent (success/error), that allows to monitor consul-templaterb with Prometheus
monitoring as well as have a look on changes on Services in your Consul Cluster.

For instance, it allows to see immediatly which service is updated the most
often with topk(100, consul_service_stats_requests_total) so performing
most writes in the Cluster.

Change-Id: I2bcced56bf7575d232aa6f1aef9e81b9f3599e7a