Fix template collision for json endpoints
With regular consul/vault endpoints, the url path + query parameters
were used as unique identifiers for an endpoint.

With as_json, we've opened the way to strange apis (non-REST, RPC-like) which
accept commands as request body on a single url. For instance: api/v1
and body contains:
{ "type": "<do stuff>" }

Before this patch, all endpoints were identified the same way and lead
to a collision.
Now we properly identify different endpoints.

Change-Id: I94c4b1237e61986af18eb6455b4ca365f696d917